chat-with-ai/requirements.txt
cohere==4.17.0
Flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0